# bin/main/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Decision Engine — Rule Platform API
  description: |
    # Decision Engine — Plataforma de Regras de PLD

    API da plataforma de decisão para Prevenção à Lavagem de Dinheiro (PLD).
    Desacopla a lógica de **detecção** (Screening Context) da lógica de **decisão**
    (quando gerar alertas, com quais critérios, baseado em quais fatos).

    ## Arquitetura

    ```
    ┌─────────────────┐     DetectionEvent      ┌──────────────────┐    DecisionMadeEvent    ┌───────────────┐
    │  Screening      │ ──────────────────────▶ │  Decision        │ ─────────────────────▶ │  Alert        │
    │  Context (MF09) │                         │  Context         │                        │  Context      │
    │                 │                         │                  │                        │               │
    │  Detecta        │                         │  Avalia regras   │                        │  Cria alertas │
    │  keywords       │                         │  Decide ações    │                        │  com status   │
    └─────────────────┘                         └──────────────────┘                        └───────────────┘
    ```

    ## Fluxo de Decisão

    ```
    1. Transação PIX recebida
    2. KeywordScreeningService detecta termo restrito → publica DetectionEvent
    3. Decision Engine consome evento:
       a) Verifica idempotência (transactionId + ruleId)
       b) Carrega configuração ativa da regra
...
    5. Publica DecisionMadeEvent
...

...

...

...

...
servers:
...
paths:
...
  /v1/decision/facts:
    get:
      tags: [Fact Registry]
      summary: Listar fatos disponíveis
      description: |
        Retorna os fatos (Facts) disponíveis para uso em configurações de regras.
        Cada fact possui tipo, operadores suportados e entidade de origem.

        O analista consulta este catálogo para saber quais facts pode usar ao criar configurações.
      operationId: listFacts
      parameters:
        - name: entity
          in: query
          required: false
          description: "Filtrar por entidade de origem (ex.: Risk, Screening, Customer)"
          schema:
            type: string
          example: Risk
        - name: enabled
          in: query
          required: false
          description: Filtrar por status habilitado/desabilitado
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de definições de fatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactDefinitionResponse'
              examples:
                mvp:
                  summary: Fatos do MVP
                  value:
                    - id: "661e8400-e29b-41d4-a716-446655440001"
                      name: "keywordMatched"
                      displayName: "Keyword Matched"
                      entity: "Screening"
                      type: "BOOLEAN"
                      context: "SCREENING"
                      source: "Screening"
                      supportedOperators: ["EQUALS", "NOT_EQUALS"]
                      enabled: true
                    - id: "661e8400-e29b-41d4-a716-446655440002"
                      name: "customerRisk"
                      displayName: "Customer Risk"
                      entity: "Risk"
                      type: "ENUM"
                      context: "SCREENING"
                      source: "PLD"
                      supportedOperators: ["EQUALS", "NOT_EQUALS", "GREATER_THAN_OR_EQUAL"]
                      enabled: true
...
  /v1/alerts:
    get:
      tags: [Alerts]
      summary: Consultar alertas
      description: |
        Consulta alertas gerados pelo Decision Engine.
        **Pelo menos um filtro é obrigatório** (transactionId ou ruleId).

        Alertas são criados automaticamente quando:
        - Decision Engine decide ALERT
        - Action configurada é GENERATE_ALERT
        - DecisionMadeEvent é publicado com sucesso

        Idempotência: apenas 1 alerta por (transactionId, ruleId).
      operationId: searchAlerts
      parameters:
        - name: transactionId
          in: query
          required: false
          description: Filtrar por ID da transação
          schema:
            type: string
          example: "TX-001"
        - name: ruleId
          in: query
          required: false
          description: Filtrar por ID da regra (paginado)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Lista de alertas ou resultado paginado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'
              examples:
                openAlert:
                  summary: Alerta aberto
                  value:
                    - id: "990e8400-e29b-41d4-a716-446655440000"
                      transactionId: "TX-001"
                      ruleId: "550e8400-e29b-41d4-a716-446655440000"
                      customerId: "CUST-42"
                      facts:
                        keywordMatched: true
                        customerRisk: "AR"
                      configurationVersion: 3
                      traceId: "trace-uuid-123"
                      actions: ["GENERATE_ALERT"]
                      explanation:
                        traceId: "trace-uuid-123"
                        decision: "ALERT"
                      status: "OPEN"
                      createdAt: "2025-01-15T10:35:00Z"
                      updatedAt: "2025-01-15T10:35:00Z"
        '400':
          description: Nenhum filtro informado
  /v1/alerts/{alertId}:
    get:
      tags: [Alerts]
      summary: Buscar alerta por ID
      operationId: getAlertById
      parameters:
        - name: alertId
          in: path
          required: true
          description: UUID do alerta
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alerta encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Alerta não encontrado
  /v1/alerts/{alertId}/status:
    patch:
      tags: [Alerts]
      summary: Atualizar status do alerta
      description: |
        Atualiza o status de um alerta seguindo a state machine:

        ```
        ┌──────┐          ┌──────────────┐          ┌────────┐
        │ OPEN │ ────────▶│ UNDER_REVIEW │ ────────▶│ CLOSED │
        │      │          │              │          │        │
        └──────┘          └──────┬───────┘          └────────┘
                                 │
                                 │          ┌────────────────┐
                                 └────────▶│ FALSE_POSITIVE │
                                            │                │
                                            └────────────────┘
        ```

        **Transições válidas:**
        - OPEN → UNDER_REVIEW
        - UNDER_REVIEW → CLOSED
        - UNDER_REVIEW → FALSE_POSITIVE

        **Estados terminais** (sem transição possível):
        - CLOSED
        - FALSE_POSITIVE
      operationId: updateAlertStatus
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlertStatusRequest'
            examples:
              startReview:
                summary: Iniciar revisão
                value:
                  status: "UNDER_REVIEW"
              closeAlert:
                summary: Fechar alerta (confirmado)
                value:
                  status: "CLOSED"
              falsePositive:
                summary: Marcar como falso positivo
                value:
                  status: "FALSE_POSITIVE"
      responses:
        '200':
          description: Status atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Alerta não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Transição de status inválida (ex.: OPEN → CLOSED, CLOSED → qualquer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidTransition:
                  value:
                    timestamp: "2025-01-15T10:40:00Z"
                    status: 422
                    error: "Unprocessable Entity"
                    message: "Transição inválida: OPEN → CLOSED"
                    details: null

  # ═══════════════════════════════════════════════════════════════════════════════
  # KEYWORD SCREENING — Detecção de termos restritos (MF09)
  # ═══════════════════════════════════════════════════════════════════════════════
...
components:
  schemas:
...
    RuleContext:
      type: string
      description: |
        Contexto ao qual uma Rule Definition pertence.
        Identifica o domínio de avaliação da regra.
      enum: [SCREENING, TRANSACTION, CUSTOMER, ACCOUNT]
      example: SCREENING
...
    AlertStatus:
      type: string
      description: |
        Status do alerta com state machine.
        OPEN → UNDER_REVIEW → CLOSED | FALSE_POSITIVE (terminais).
      enum: [OPEN, UNDER_REVIEW, CLOSED, FALSE_POSITIVE]
      example: OPEN
    ComparisonOperator:
      type: string
      description: |
        Operador de comparação para expressões.
        **MVP suporta:** EQUALS, NOT_EQUALS, GREATER_THAN_OR_EQUAL.
        Outros operadores estão modelados para evolução futura.
      enum: [EQUALS, NOT_EQUALS, GREATER_THAN, GREATER_THAN_OR_EQUAL, LESS_THAN, LESS_THAN_OR_EQUAL, IN, NOT_IN, CONTAINS]
      example: EQUALS
...
    FactType:
      type: string
      description: |
        Tipo do valor de um Fact.
        Determina quais operadores são válidos e como o valor é interpretado.
      enum: [BOOLEAN, ENUM, MONEY, STRING, NUMBER]
      example: BOOLEAN

    # ═══════════════════════════════════════════════════════════════════════════════
    # REQUEST DTOs
    # ═══════════════════════════════════════════════════════════════════════════════
...
    UpdateAlertStatusRequest:
      type: object
      description: Request para transicionar o status de um alerta.
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/AlertStatus'
          description: |
            Novo status desejado. Deve respeitar a state machine:
            OPEN → UNDER_REVIEW → CLOSED | FALSE_POSITIVE

    # ═══════════════════════════════════════════════════════════════════════════════
    # RESPONSE DTOs
    # ═══════════════════════════════════════════════════════════════════════════════
...
    FactDefinitionResponse:
      type: object
      description: |
        Definição de um Fact no catálogo.
        Descreve um dado contextual que pode ser usado em expressões de regras.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do fact
        name:
          type: string
          description: Nome técnico do fact (usado em expressions)
          example: "customerRisk"
        displayName:
          type: string
          description: Nome amigável para exibição na UI
          example: "Customer Risk"
        entity:
          type: string
          description: Entidade de negócio à qual o fact pertence
          example: "Risk"
        type:
          $ref: '#/components/schemas/FactType'
          description: |
            Tipo do valor. Determina qual tipo de `expectedValue` usar nas expressions:
            - BOOLEAN → true/false
            - ENUM → String (ex.: "BR", "MR", "AR")
            - NUMBER → Numérico
            - STRING → Texto livre
            - MONEY → Objeto {amount, currency}
        context:
          $ref: '#/components/schemas/RuleContext'
          description: Contexto onde o fact é utilizado
        source:
          type: string
          description: Bounded context de origem dos dados
          example: "PLD"
        supportedOperators:
          type: array
          description: |
            Operadores de comparação que este fact suporta.
            O analista só pode usar estes operadores ao criar expressões com este fact.
          items:
            $ref: '#/components/schemas/ComparisonOperator'
          example: ["EQUALS", "NOT_EQUALS", "GREATER_THAN_OR_EQUAL"]
        enabled:
          type: boolean
          description: Se o fact está habilitado para uso em novas configurações
          example: true
...
    AlertResponse:
      type: object
      description: |
        Alerta gerado automaticamente pelo Decision Engine.
        Criado quando decision=ALERT e action=GENERATE_ALERT.
        Idempotente: apenas 1 alerta por (transactionId, ruleId).
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do alerta
        transactionId:
          type: string
          description: ID da transação que originou o alerta
          example: "TX-001"
        ruleId:
          type: string
          format: uuid
          description: ID da regra que gerou o alerta
        customerId:
          type: string
          description: ID do cliente associado à transação
          example: "CUST-42"
        facts:
          type: object
          description: Fatos avaliados que levaram ao alerta
          additionalProperties: true
          nullable: true
          example:
            keywordMatched: true
            customerRisk: "AR"
        configurationVersion:
          type: integer
          description: Versão da configuração que estava ativa
          nullable: true
          example: 3
        traceId:
          type: string
          description: TraceId para correlação com a DecisionExecution
          nullable: true
          example: "trace-uuid-123"
        actions:
          type: array
          description: Ações que foram executadas
          items:
            type: string
          nullable: true
          example: ["GENERATE_ALERT"]
        explanation:
          type: object
          description: Dados resumidos da explicação da decisão
          additionalProperties: true
          nullable: true
        status:
          $ref: '#/components/schemas/AlertStatus'
          description: Status atual do alerta (segue state machine)
        createdAt:
          type: string
          format: date-time
          description: Quando o alerta foi criado
          example: "2025-01-15T10:35:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Última atualização de status
          example: "2025-01-15T10:35:00Z"

    # ═══════════════════════════════════════════════════════════════════════════════
    # SCREENING CONTEXT — Keyword & Contextual Screening
    # ═══════════════════════════════════════════════════════════════════════════════
...
    ErrorResponse:
      type: object
      description: |
        Formato padronizado de erro para todos os endpoints.
        Inclui timestamp, status HTTP, código de erro e mensagem descritiva.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento do erro (ISO-8601)
          example: "2025-01-15T10:30:00Z"
        status:
          type: integer
          description: Código HTTP do erro
          example: 422
        error:
          type: string
          description: Descrição curta do tipo de erro HTTP
          example: "Unprocessable Entity"
        message:
          type: string
          description: Mensagem detalhada do erro (legível)
          example: "Fato 'unknownFact' não encontrado no catálogo"
        details:
          type: string
          description: "Detalhes adicionais (ex.: campos com erro de validação)"
          nullable: true
          example: "expressions[0].factName: não pode ser vazio; createdBy: é obrigatório"